chore(cargo-near): move  `src/common.rs` -> `src/types/color_preference.rs`
