+++ cargo-near/src/types/color_preference.rs
