+++ cargo-near/src/types/mod.rs
@@ -1,3 +1,4 @@
+pub mod color_preference;
 pub mod metadata;
 pub mod source_id;
 pub mod utf8_path_buf;
